requirement.txt
asgiref==3.8.1
dj-database-url==2.2.0
Django==5.0.7
djangorestframework==3.15.2
gunicorn==22.0.0
packaging==24.1
psycopg2==2.9.9
python-dotenv==1.0.1
sqlparse==0.5.0
typing_extensions==4.12.2
